Persist `capabilities` tag vocabulary
We're now providing a `capabilities_selected` list to the
package_metadata_fields.html form, which means capabilities added will
both persist in the datastore but also be repopulated in the form.

A huge time-sink in this was that documentation makes it look like you
can retrieve tags by vocabulary_name e.g.

`"capabilities_selected": ("convert_from_tags")('capabilities')`

but this doesn't appear to work (either a bug or my misread of the docs)

Instead we call up the id of the vocab to use displaying selected
capabilities in the form.
```
capability_vocab_id = toolkit.get_action("vocabulary_show")(
     {}, {"id": CAPABILITY_VOCAB}
)["id"]

"capabilities_selected": [
     toolkit.get_converter("convert_from_tags")(capability_vocab_id),
]
```